==> P2_Matrix.Rmd <==
---
title: "P2 Matrix"
author: "Olive Kirk"
date: "07/02/2023"
output: pdf_document
---

Compares the overlap of genes between genesets. Also preliminary distance matrix and clustering graphs.


# Imports
```{r, echo=FALSE, warning=FALSE, message=FALSE}
# Import libraries.

knitr::opts_chunk$set(echo = TRUE)

# Core computational packages.
library(clusterProfiler)
library(stats)
library(matrixStats)
library(org.Mm.eg.db)
library(EnrichmentBrowser)
library(rWikiPathways)
library(cluster)

# Visualisation packages.
library(ggplot2)
library(ggfortify)
library(ComplexHeatmap)
library(paletteer)
library(factoextra)

# Coding preference packages.
library(readr)
library(tibble)
library(data.table)

set.seed(42)
options("scipen"=10, "digits"=4)
time_date <- paste0(Sys.Date(), "--", format(Sys.time(), "%H-%M-%S--"))
dev.new(width=10, height=10, unit="in")

folder = ""
#folder = "_up_reg"
#folder = "_down_reg"

```

```{r}

# Define functions.

# Transpose matrix and retain the name values.
t_pose <- function(d) {
  df_t <- transpose(d)
  rownames(df_t) <- colnames(d)
  colnames(df_t) <- rownames(d)
  d <- df_t
  
  return(d)
}

```


## db imports

```{r}

get_wikipathways <- function(species="Mus musculus") {
  
  # Get metadata.
  pw <- listPathways(species)
  
  # Compile pathways.
  gs <- list()
  for (i in pw$id)
  {
    #gs <- list(unlist(gs), getXrefList(i, "L"))
    #print(i)
    gs[[length(gs)+1]] <- getXrefList(i, "L")
  }
  
#  gs <- list(unlist(gs))
  
  # Assign names.
  names(gs) <- pw$name
  
  return(gs)
}

```

```{r}

get_reactome <- function(){
  
  cpdb <- read.csv("data/CPDB_pathways_genes.csv", sep="\t")
  db <- cpdb[cpdb$source == "Reactome",]
  
  # Compile pathways.
  gs <- list()
  for (i in 1:dim(db)[1])
  {
    gs[[length(gs)+1]] <- unlist(strsplit(db$entrez_gene_ids[i], ","))
  }
  
  # Assign names.
  names(gs) <- db$pathway
  
  return(gs)
}

```

```{r}

get_mousecyc <- function(){
  
  cpdb <- read.csv("data/CPDB_pathways_genes.csv", sep="\t")
  db <- cpdb[cpdb$source == "MouseCyc",]
  
  # Compile pathways.
  gs <- list()
  for (i in 1:dim(db)[1])
  {
    gs[[length(gs)+1]] <- unlist(strsplit(db$entrez_gene_ids[i], ","))
  }
  
  # Assign names.
  names(gs) <- db$pathway
  gs <- gs[-202]
  
  return(gs)
}

```


# Set up

```{r}
# Choose which pathway database to use.

db <- "kegg"
#db <- "go"
db <- "msigdb-h"
#db <- "wikipathways"
#db <- "reactome"
#db <- "mousecyc"

if (db == "kegg") mmu.gs <- getGenesets(org="mmu", db="kegg")
if (db == "go") mmu.gs <- getGenesets(org="mmu", db="go", onto="BP")
if (db == "msigdb-h") mmu.gs <- getGenesets(org="mmu", db="msigdb", cat="H")
if (db == "wikipathways") mmu.gs <- get_wikipathways()
if (db == "reactome") mmu.gs <- get_reactome()
if (db == "mousecyc") mmu.gs <- get_mousecyc()

print(length(mmu.gs))

```

```{r}

distance_matrix <- function(sigsets){

  # Create an overlaps matrix the quantify the overlapping sets.  
  overlap.matrix <- matrix(, nrow=length(sigsets), ncol=length(sigsets))
  
  for (i in c(1:length(sigsets)))
  {
    for (j in c(i:length(sigsets)))
    {
      # Jaccard Index.
      overlap.matrix[i,j] <- 1 - ((length(comparelists(sigsets[[i]], sigsets[[j]])$intersect)) / length(unique(c(sigsets[[i]], sigsets[[j]]))))
      overlap.matrix[j,i] <- overlap.matrix[i,j]
    }
  }
  
  # for (i in c(1:length(sigsets)))
  # {
  #   for (j in c(1:length(sigsets)))
  #   {
  #     # Jaccard Index.
  #     overlap.matrix[i,j] <- (length(comparelists(sigsets[[i]], sigsets[[j]])$intersect)) / length(unique(c(sigsets[[i]], sigsets[[j]])))
  #   }
  # }
  
  colnames(overlap.matrix) <- names(sigsets)
  rownames(overlap.matrix) <- names(sigsets)
  
  #overlap.matrix <- as.dist(overlap.matrix)
  
  # Return for clustering.
  return(overlap.matrix)
}

```

```{r}
temp <- distance_matrix(mmu.gs)

# Visualise the results.
if (db != "go")
{
  Heatmap(temp, #row_names_gp = gpar(fontsize = 25), column_names_gp = gpar(fontsize = 25),
        cluster_rows=FALSE, cluster_columns=FALSE, row_dend_reorder=FALSE, column_dend_reorder=FALSE,
        column_names_side="top", column_names_rot=-90, show_heatmap_legend = FALSE)  
  
  #fviz_dist(as.dist(temp))
}

```

```{r}
# CLuster results to find relevant overlap cutoff point.
# source: https://www.statmethods.net/advstats/cluster.html

mydata <- unlist(as.list(temp[-lower.tri(temp)])) # make one list of metrics.
print(length(mydata))
mydata <- na.omit(mydata) # listwise deletion of missing
#mydata <- scale(mydata) # standardize variables 
print(length(mydata))
```


```{r}
# K-Means Cluster Analysis

fit <- kmeans(mydata, 3) 
# get cluster means
aggregate(mydata,by=list(fit$cluster),FUN=mean)
# append cluster assignment
#data.frame(mydata, fit$cluster) 

# KEGG
# three clusters gives 0.00 / 0.99 / 0.12
# four clusters gives 0.00 / 0.99/ 0.078/ 0.200
# --> take 0.08 / 8% as the cut off value

# GO

# Reactome
# three clusters gives 0.00/ 0.17/ 0.78
# four clusters gives 0.00 / 0.88/ 0.10/ 0.36
# five clusters gives 0.00 / 0.92/ 0.06/ 0.19/ 0.45
# --> take 0.08 / 8% as the cut off value

# MouseCyc
# three clusters gives 0.00 / 0.99 / 0.12
# four clusters gives 0.00 / 0.99/ 0.065/ 0.248
# --> take 0.08 / 8% as the cut off value


p <- 0.08 # KEGG, Reactome, MouseCyc

```

```{r}

hc.res <- hclust(as.dist(temp), method="average")
par(mar = c(15, 2, 1, 0))
#plot(hc.res, labels=FALSE)
plot(as.dendrogram(hc.res))
#plot(hc.res, hang=-1)
# plot(hc.res, labels = FALSE,  hang=-1)
#fviz_dend(hc.res, k=10, palette="jco", type="phylogenic")
#fviz_dend(hc.res, type="phylogenic", phylo_layout="layout_as_tree")

```

```{r}

#hc.res
#get_clust_tendency(temp, 49)

```



# PW search

```{r}
stop("Examination section ahead--")
```


```{r}

# Select the name of the pathway of interest.

# MouseCyc.
n <- "folate polyglutamylation"
n <- "UDP-<i>N</i>-acetyl-D-glucosamine biosynthesis II"
n <- "UDP-galactose biosynthesis (salvage pathway from galactose using UDP-glucose)"
n <- "cysteine biosynthesis III (mammalia)"
n <- "sphingomyelin metabolism"

# KEGG.
n <- "mmu05417_Lipid_and_atherosclerosis"
n <- "mmu04020_Calcium_signaling_pathway"
n <- "mmu04710_Circadian_rhythm"
n <- "mmu00750_Vitamin_B6_metabolism"
n <- "mmu04550_Signaling_pathways_regulating_pluripotency_of_stem_cells"
n <- "mmu04360_Axon_guidance"

# WikiPathways.
n <- "Lung fibrosis"
n <- "Spinal cord injury"
n <- "Ethanol metabolism resulting in production of ROS by CYP2E1"
n <- "ID signaling pathway"
n <- "Fatty acid oxidation"
n <- "Hypertrophy model"
n <- "ESC pluripotency pathways"
n <- "Delta-Notch signaling pathway"
n <- "Adar1 editing defficiency immune response"

# Reactome.
n <- "RHO GTPases Activate ROCKs"
n <- "NIK-->noncanonical NF-kB signaling"
n <- "Activation of NF-kappaB in B cells"
n <- "Growth hormone receptor signaling"
n <- "RAF/MAP kinase cascade"
n <- "Voltage gated Potassium channels"
n <- "ABC transporters in lipid homeostasis"
n <- "PTK6 Regulates RHO GTPases, RAS GTPase and MAP kinases"
n <- "VEGFA-VEGFR2 Pathway"
n <- "ERBB2 Regulates Cell Motility"
n <- "ERK/MAPK targets"
n <- "VEGFR2 mediated cell proliferation"
n <- "RUNX2 regulates bone development"
n <- "Signaling by Receptor Tyrosine Kinases"
n <- "PTK6 Regulates RHO GTPases, RAS GTPase and MAP kinases"
n <- "Gap junction trafficking and regulation"
n <- "Gap junction trafficking"
n <- "Gap junction degradation"
n <- "Death Receptor Signalling"
n <- "Interferon Signaling"
n <- "Caspase activation via extrinsic apoptotic signalling pathway"
# n <- "DEx/H-box helicases activate type I IFN and inflammatory cytokines production"
n <- "TP53 Regulates Transcription of Genes Involved in Cytochrome C Release"
n <- "Activation of NIMA Kinases NEK9, NEK6, NEK7"
n <- "Cell surface interactions at the vascular wall"
n <- "ERK/MAPK targets"
n <- "Cell junction organization"
n <- "mitochondrial fatty acid beta-oxidation of saturated fatty acids"

# GO.
n <- "GO:0001649_osteoblast_differentiation"
n <- "GO:0007422_peripheral_nervous_system_development"
n <- "GO:0035914_skeletal_muscle_cell_differentiation"
n <- "GO:0060426_lung_vasculature_development"
n <- "GO:0014911_positive_regulation_of_smooth_muscle_cell_migration"
n <- "GO:0032966_negative_regulation_of_collagen_biosynthetic_process"
n <- "GO:0071222_cellular_response_to_lipopolysaccharide"
n <- "GO:0006119_oxidative_phosphorylation"
n <- "GO:0061419_positive_regulation_of_transcription_from_RNA_polymerase_II_promoter_in_response_to_hypoxia" 
n <- "GO:0090050_positive_regulation_of_cell_migration_involved_in_sprouting_angiogenesis"
n <- "GO:0048514_blood_vessel_morphogenesis"
n <- "GO:0007288_sperm_axoneme_assembly" 
n <- "GO:0060425_lung_morphogenesis"   
n <- "GO:0048701_embryonic_cranial_skeleton_morphogenesis"  
n <- "GO:0008016_regulation_of_heart_contraction" 
n <- "GO:0035924_cellular_response_to_vascular_endothelial_growth_factor_stimulus"
n <- "GO:0030500_regulation_of_bone_mineralization" 
n <- "GO:0055013_cardiac_muscle_cell_development"
n <- "GO:0008585_female_gonad_development"
n <- "GO:0007140_male_meiotic_nuclear_division"
n <- "GO:0038033_positive_regulation_of_endothelial_cell_chemotaxis_by_VEGF-activated_vascular_endothelial_growth_factor_receptor_signaling_pathway"
n <- "GO:0060326_cell_chemotaxis"
n <- "GO:0002051_osteoblast_fate_commitment"
n <- "GO:0032496_response_to_lipopolysaccharide"
n <- "GO:0031103_axon_regeneration"
n <- "GO:0006814_sodium_ion_transport"    
n <- "GO:0006002_fructose_6-phosphate_metabolic_process"
n <- "GO:0033499_galactose_catabolic_process_via_UDP-galactose"
n <- "GO:0042149_cellular_response_to_glucose_starvation"


# Print the list of noticably overlapping pathways.
m <- which(colnames(temp) %in% n)
print(paste0("[", db,"] :: ", n))
print(as.data.frame(temp[m, temp[m,] > p]))
  
```
